+++ addons/csc/functions/accessPoint/fn_createAccessPoint.sqf
@@ -23,6 +23,8 @@ params [
     ["_addParams",        createHashMap, [createHashMap]]
 ];
 
+if (isNull _targetObject) exitWith {};
+
 switch (true) do {
     case (_crates isEqualTo "DEFAULT"): { _crates = ["DEFAULT", "CRATES"] call FUNC(getDefaultPresets); };
     case (_crates isEqualType ""): { _crates = [_crates]; };
@@ -61,21 +63,7 @@ private _conditionCode = switch (_accessPoint getOrDefault ["conditionType", "AL
     default { { true } };
 };
 
-
-private _aceAction = [
-    QGVAR(AceAction)                                            // * 0: Action name <STRING>
-    ,"Request Custom Supply Crates"                             //  * 1: Name of the action shown in the menu <STRING>
-    ,"\a3\ui_f\data\IGUI\Cfg\simpleTasks\types\box_ca.paa"      //  * 2: Icon <STRING> "\A3\ui_f\data\igui\cfg\simpleTasks\types\backpack_ca.paa"
-    ,FUNC(openDialog)                                           //  * 3: Statement <CODE>
-    ,_conditionCode                                                      //  * 4: Condition <CODE>
-    ,{}                                                         //  * 5: Insert children code <CODE> (Optional)
-    ,_accessPoint                                               //  * 6: Action parameters <ANY> (Optional)
-//    ,[0,0,0]                              //  * 7: Position (Position array, Position code or Selection Name) <ARRAY>, <CODE> or <STRING> (Optional)
-//    ,20                                   //  * 8: Distance <NUMBER> (Optional)
-//    ,[false,false,false,false,false]      //  * 9: Other parameters [showDisabled,enableInside,canCollapse,runOnHover,doNotCheckLOS] <ARRAY> (Optional)
-//    ,{}                                   //  * 10: Modifier function <CODE> (Optional)
-] call ace_interact_menu_fnc_createAction;
-
+private _aceAction = [_conditionCode, _accessPoint] call FUNC(createAction);
 
 [
     _targetObject                    	// * 0: Object the action should be assigned to <OBJECT>
@@ -85,6 +73,3 @@ private _aceAction = [
 ] call ace_interact_menu_fnc_addActionToObject;
 
 ZRN_LOG_MSG_1(AccessPoint Established on,_targetObject);
-ZRN_LOG_MSG_1(Crates:,_crates);
-ZRN_LOG_MSG_1(Destinations:,_destinations);
-ZRN_LOG_MSG_1(Delivery_Modes:,_delivery_modes);
