+++ addons/csc/functions/ui/fn_ui_crates_onLBSelChanged.sqf
@@ -15,7 +15,10 @@
 * Public: No
 */
 
-params ["_ctrl", "_selectedRow"];
+params ["_ctrl", "_index"];
 
-call FUNC(ui_crates_check_amount); // update the + and - button 
+diag_log format ['[CVO](debug)(fn_ui_crates_onLBSelChanged) _this: %1', _this];
 
+call FUNC(ui_update_arrows); // update the + and - button 
+
+call FUNC(ui_update_crate_desc);
