+++ addons/csc/functions/ui/fn_ui_update_arrows.sqf
@@ -20,8 +20,6 @@ private _display = findDisplay CVO_IDD_CSC_REQUEST;
 
 private _ctrlList = _display displayCtrl  CVO_IDC_CSC_Crates_ListNBox;
 
-// _ctrlList call cvo_common_fnc_ui_lnb_output;
-
 private _curRow = lnbCurSelRow _ctrlList;
 
 
@@ -43,21 +41,6 @@ switch (true) do {
     case (_curTotal <  _maxCrates): { ctrlEnable [CVO_IDC_CSC_Crates_ListNBox_arrowPlus, true ]; };
 };
 
-
-// Disable OK Button if more Crates are selected then maximum possible
-switch (true) do {
-    case ( _curTotal isEqualTo 0   ): {
-        ctrlEnable [CVO_IDC_CSC_ButtonOK, false];
-            ctrlSetText [CVO_IDC_CSC_Status, "No crates selected."];
-    };
-    case ( _curTotal >  _maxCrates ): {
-        ctrlEnable [CVO_IDC_CSC_ButtonOK, false];
-        ctrlSetText [CVO_IDC_CSC_Status, "Too many crates selected."];
-    };
-    case ( _curTotal <= _maxCrates ): {
-        ctrlEnable [CVO_IDC_CSC_ButtonOK, true ];
-        ctrlSetText [CVO_IDC_CSC_Status, "Request can be processed."];
-    };
-};
-
 ctrlSetText [CVO_IDC_CSC_Crates_Subtitle_Text, format ["Custom Supply Crates [ %1 / %2 ]", _curTotal, _maxCrates ]];
+
+call FUNC(ui_update_canRequest);
