+++ addons/csc/presets/cvo_csc_delivery_modes_base.hpp
@@ -1,0 +1,14 @@
+class base {
+
+    displayName = "";
+    code_description = "";
+
+    maxCrates = 1;
+
+    code = "";
+
+    scope = 2;
+
+    class parameters {
+    };
+};
