+++ addons/csc/presets/cvo_csc_destinations_base.hpp
@@ -1,0 +1,11 @@
+class base {
+
+    displayName = "";
+    description = "";
+
+    code = "";
+
+    scope = 2;
+
+    class parameters {};
+};
