+++ addons/main/script_version.hpp
@@ -1,7 +1,7 @@
 #define MAJOR 1
 #define MINOR 6
 #define PATCH 0
-#define BUILD 555
+#define BUILD 580
 
 
 // #define VERSION MACROS
